core: fix search on fedora 21
The problem was JDK 8 that is installed by default on f21
The search mechanism relies on column order retrieved from a map, it
seems that there are differences that cause that return the columns in
different order in JDK 8.
Code was changed to use a alphabetic order of the column names to build
the condition and relevant test expectations were updated accordingly.

Change-Id: I4e6f2cb0832dc21be9fab9f24858e6f7c4c613ba
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1163400
Signed-off-by: emesika <emesika@redhat.com>