packaging: setup: clearing only zombie tasks
We cannot upgrade while async tasks are still running and we have
to wait till they ends.
To avoid getting stuck we need to clear zombie task from the list
before checking if we have to wait.
If installed version is 3.5 or upper, before removing zombie tasks
we have also to remove zombie commands so we have to call
the taskcleaner utility twice.

Change-Id: I73dcb9ee0a465ef9552a085e59ab36576ee3b660
Bug-Url: https://bugzilla.redhat.com/1196136
Signed-off-by: Simone Tiraboschi <stirabos@redhat.com>