core: find connections with same password
When saving connections to the db the password is being encrypted
(introduced in change d0ed215) - the problem with that is that if we
attempt to check if a connection with specific details was already saved
- we'll not be able to fetch it (as the encrypted password is different
each time). This issue was handled in I997b by removing the password out
of the query but further fix is required, for example if connection with
wrong password already exists in the db, we'll use it instead of using
the other password.

In this change i changed the way that we check if similar connection
exists, the conenctions with the 'same' details (besides the password)
are being loaded and then it's being checked if the password is the same
as the entered one - if it is, it's the same connection and otherwise
it's a different one.

Change-Id: I0fc70dd4e7988f4ea4597f016c9de8419c17adfa
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1176402
Signed-off-by: Liron Aravot <laravot@redhat.com>