core: fix NPE on addition of external vm
We add externally managed VMs using AddVmFromScratch command. This
command is called by the monitoring thread so no user is set as the
'current user' when it is executed. Thus we got NPE when trying to set
the its disk permissions. However, this kind of VMs are diskless VMs so
we are not supposed to set disk permissions for it.

Thus, changing AddVmCommand#addDiskPermissions in a way that the disk
owners will not be retrieved for diskless VMs prevents the NPE.

Change-Id: I2e0cff201b1c8c89682fc2c3a4ddbc88b4ec803e
Bug-Url: https://bugzilla.redhat.com/1176528
Signed-off-by: Arik Hadas <ahadas@redhat.com>