webadmin: edit 'instance type' on double click
The method getEditCommand in a model must return the command that opens
the edit dialog in order to open the edit dialog on double click on one
of its items.

The model for instance-types didn't override this method, thus it
returned null and the edit dialog wasn't opened on double click on
instance type. The solution is to override the getEditCommand method and
delegate to the getEditInstanceTypeCommand method.

Change-Id: Ic15aff5ee200b9e5a9f04a397a6571f362b90c38
Bug-Url: https://bugzilla.redhat.com/1174655
Signed-off-by: Arik Hadas <ahadas@redhat.com>