engine: Send net_ifacename to VDSM when connecting iSCSI server
In order to properly configure iSCSI multipathing each iSCSI iface
should be related to the real network interface. Software iSCSI knows
how to relate iface to the network interface when all network interfaces
on that host reside on different subnets see [1].

However if multiple NICs reside on the same subnet iSCSI iface
configuration must contain a "iface.net_ifacename" representing a name
of the network layer interface, in case this network is bridged then it
should be a name of the bridge.

[1] http://www.open-iscsi.org/docs/README section 5.1.1.

Change-Id: I3e2f3549fa08741f719eeb8d70209ab298e09da3
Bug-Url: https://bugzilla.redhat.com/1178838
Signed-off-by: Sergey Gotliv <sgotliv@redhat.com>
Signed-off-by: Maor Lipchuk <mlipchuk@redhat.com>