core: Add a CDA on attach to an uninitialized Data Center
Adding a validation message when trying to attach a Storage Domain which
already have a storage pool id initiliazed in its meta data.

Storage Domain which have storage pool id already initialized in its
meta data (For example after disaster), needs to be detached first and
then to be attached to the relevant Data Center, since the detach
operation is an SPM operation, an un-initialized Data Center can't
perform this operation since it does not have an SPM yet.

Change-Id: I13ec66cc7ff4773221cc994cb4f0be1886ed9a6b
Bug-Url: https://bugzilla.redhat.com/1138126
Signed-off-by: Maor Lipchuk <mlipchuk@redhat.com>