core: Add vdc query for import Block Storage Domain
Add vdc query for validating that the block Storage Domain is not
attached to another Data Center.
The validation is being done on the Storage Domain metadata which
contains the storage pool id.

The Query only fetches the storage domain info and does not perform
connect or disconnect, to prevent a scenario of disconnecting an exisitng
Storage Domains.

Change-Id: Id93a96ad7fdad3305f3f2254f67f476c1e989b6f
Bug-Url: https://bugzilla.redhat.com/1176552
Signed-off-by: Maor Lipchuk <mlipchuk@redhat.com>