core: model hibernation volumes as disks in database
Hibernation volumes were represented as comma-separated string:
<storage domain>,<storage pool>,<dump disk id>,<dump image id>,
<conf disk id>,<conf image id>

This patch changes the memory dump and memory conf volumes to
be modeled as disks in the database. We distinguish memory
related disks from others by checking if the snapshots table
refers to them as such that contain memory dump or memory conf.

We currently filter all the memory related disks in most of the
queries from the database so they are not displayed in the grid
of disks for example. We expose them only in queries that must
return them in order to be able to restore hibernated VM and to
delete them when they are not relevant anymore. The new views
are temporary!

This patch is only a step towards removing the comma-separated
string representation of the memory volumes and to treat them
as regular disks in different flows in the system.

In addition, this patch converts hibernation volumes that
existed before this change to disks.

Change-Id: I295bd98ead191c1eab54c471ea28a217dee52051
Signed-off-by: Arik Hadas <ahadas@redhat.com>