core: store hibernation volumes in active snapshots
Change hibernate command to save the memory volumes representation in
the active snapshot instead of in vm_dynamic, and all references to the
hibernation volumes in vm_dynamic are changed accordingly.

That way hibernation volumes are stored as memory which is part of
snapshot is stored and that will help us convert all the memory images
to disks.

Change-Id: I30890eb1a75132aa4f67f754c357edc847819908
Signed-off-by: Arik Hadas <ahadas@redhat.com>