userportal/webadmin: fix missing fields in cloner
Added fields which were missing in the cloneVM part.

The 'edit template-version' feature bypass part of the saving process of
a running edited vm with a different template version(code now ignores
the model data in this scenario because it represents the new
template data which won't be relevant till shutdown). This causes the
saving process to be based solely on VmListModel::preSave() which uses this
cloner as it's only source of vm data.

Change-Id: Id475341c0cb97612927627b0bd5dfca070e19d22
Bug-Url: https://bugzilla.redhat.com/1140569
Signed-off-by: Eldan Shachar <eshachar@redhat.com>