core: Add vdc query for import File Storage Domain
Add vdc query for validating that the file Storage Domain is not
attached
to another Data Center.
The validation is being done on the Storage Domain metadata which
contains the storage pool id.

The Query only fetches the storage domain info and does not perform
connect or disconnect, to prevent a scenario of disconnecting exisitng
Storage Domains.

Change-Id: I63a90b86b8427bd6128becccf83be7204637e611
Bug-Url: https://bugzilla.redhat.com/1176552
Signed-off-by: Maor Lipchuk <mlipchuk@redhat.com>