core: ActivateDeactivateSingleAsyncOperationFactory- wrong type
ActivateDeactivateSingleAsyncOperationFactory expects the recieved VDS
list to be an ArrayList. That list is usually passed as returned by the
DAO, which means that the concrete type depends on the spring jdbc
internal implementation.

In this patch I create an ArrayList out of the passed List, I didn't
change the type as the entire "Operation" infrastracture use that list
for different operation by issuing get() (I assume that this is the
reason ArrayList was used in the first place). We need to make further
changes here, but this is a quick fixup.

As that infrastracture isn't used so commonly, this solution seems good
enough in the meanwhile.

Change-Id: If0de8b0f5838e7a464cf093b2deeeda67b32a560
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1177595
Signed-off-by: Liron Aravot <laravot@redhat.com>