core: Clean up query test conf mocking
Previously, AbstractQueryTest defined a @ClassRule for mocking Config
values. Since @ClassRules are hidden, not inherited, if any descending
class wished to have it's own MockConfigRule, the mocking from the base
class had to be copied to the descending class in order not to lose it.
This approach is both error prone and conceptually lacking, as it forces
the descendant classes to know details about the parent's
implementation.

This patch attempts to improve the situation by have AbstractQueryTest
define the getExtraConfigDescriptors() method. This allows each
inheriting class to specify which Config values it wishes to add to the
mocking, while leaving the actual mocking to the base class.

Change-Id: I1d511b1f67e842a64b7c9322a0d526e717f4a601
Signed-off-by: Allon Mureinik <amureini@redhat.com>