webadmin: MoveOrCopyDiskModel: postInitStorageDomains refactoring.
postInitStorageDomains function was using multiple filtering on the same
list of storage domains to create destination storage domain list for
each disk model in this model. Going through this list again and again
is redundant, thus this patch changes it to makes a single run on the
domains list and builds the destination array in one iteration.

Change-Id: I7538a341c3a7059b0a6ee369ca1a9b33648df456
Signed-off-by: Amit Aviram <aaviram@redhat.com>