engine: Modified fk_validator to handle composite foreign keys
In case of composite foreign keys, the table_col and fk_col in
the select references cursor was empty, which causes the validation
to error out.
for instance, if fk is on a composite primary key like this
"(sched_name, trigger_name, trigger_group)"
the regular expression should consider "," and space.

Also,in the query to fetch violations, the inner select
clause for composite keys should return multiple columns as
opposed to single column - hence trimming "()"

Change-Id: Ie4210537c6ef7ce7f89bd73be9e44178b6b8d43a
Signed-off-by: Sahina Bose <sabose@redhat.com>