restapi: Discard blank "from" parameter
Currently when the RESTAPI receives a query with an empth "from"
parameter it generates an exception because it tries to parse the value
as a numeric value. Well behaved callers shouldn't generate this kind of
request, but anyhow the RESTAPI shouldn't fail in this case, it should
just ignore the parameter completely.

Change-Id: I1a909193d17503bb7eb7025b40a56243c6a044bf
Bug-Url: https://bugzilla.redhat.com/1179171
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>