engine: fix graphics devices handling when creating pool
This patch fixes the problem when creating vm pool via UI. Previously, graphics
types of created pool weren't reset in the ui code which caused possible
inheriting of graphics device from a template.

Also, the way of storing graphics devices in AddVmPoolWithVmsParams
was wrong (it wasn't possible to tell the command that particular
graphics device was "reset" in params). This patch changes it so that
this class behaves in the same way as Vm/Template management classes.

Change-Id: I95a6208490709d263ab1bd6e4d5eb9cc6ecb769f
Signed-off-by: Frantisek Kobzik <fkobzik@redhat.com>
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1177231