webadmin: add next run configuration for pooled vms
If the pool is based on an instance type and the instance type changes, the vms
in this pool changes as well. If such a VM is running, the "next run" icon has
to be shown next to the VM type icon.

This has already been working on not-pooled vms and in user portal - this adds
the last missine part which is the pooled vms in web admin.

Change-Id: Iaf53f83111a84065cf734612541a6144d90fc3ef
Bug-Url: https://bugzilla.redhat.com/1176937
Signed-off-by: Tomas Jelinek <tjelinek@redhat.com>