aaa: Modify webadmin and userportal to use enginesso for authentication
Add webadmin side code to support authentication using
engine sso. Modify webadmin and user portal to use the
new SSO filters for authorization.

Session validation on engine side is done using
sso

Modify the welcome page to show the current
user logged in and provide a link to switch
user by invalidating the current session and
turning off external auth

Basic auth and negotiate filters on engine side
are not required

SSO token and JSESSION from rest api should be
obtained using Engine Session Id and SSO Session Id.

The setup is sso aware and registers the engine with the
sso as a client.

Remove code that performs login in UI.

Login user and admin commands and the parameter
can be deleted from the backend as the login sequence
is handled by SSO

Change-Id: Iff0aee9d0f5ee606ff7f397cab69017ca7d9df08
Bug-Url: https://bugzilla.redhat.com/1092744
Signed-off-by: Ravi Nori <rnori@redhat.com>