Override firewall configurations on deploy for provisioned hosts
Currently we don't have parameter for override-firewall configurations
in hosts\[id]\install restapi. Therefore we cannot pass this parameter
along
the way from foreman to engine. Until adding the parameter to the API
this patch hacks the code to always override the firewall configurations
for vdsm deploy for each foreman installation when updateVdsCommand is
initiated by the reinstall request. This is monitored by checking that
the host in InstallingOS status, which can be only if the host is
firstly being provisioned by foreman.

https://bugzilla.redhat.com/show_bug.cgi?id=1177126
Signed-off-by: Yaniv Bronhaim <ybronhei@redhat.com>
Change-Id: I2634c33d87e9ef8ef41deeac28287b3b7ae1344a