core: adjusts default_display_type and video card for blank template
This patch sets default_display_type of vm/template so that it corresponds to
video device of this vm/template. This change is needed since DisplayType enum
(and its semantics) was changed.

It also changes the video device of blank template. Previously the
blank template had 'cirrus' video device, but it had
default_display_type set to 1 (which was 'qxl' display type). This
patch changes this inconsistence and sets both values
(default_display_type and vm_device type) to values corresponding to
qxl.

Change-Id: I63b268619065fef638349d95bef22bc0ce34c9b8
Signed-off-by: Frantisek Kobzik <fkobzik@redhat.com>