packaging: add gwt symbols packages
two additional packages:
 * ovirt-engine-userportal-debuginfo
 * ovirt-engine-webadmin-portal-debuginfo

these packages contain the gwt symbols required to interpret stack
trace.

separate packages as these files are huge and in most cases are not
required.

also if we add a feature to decipher stack trace on the fly, we can ask
a customer to install these packages so they will be available for
engine.

Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1179862
Change-Id: I0c5fe0de24a83e2ce49c27f6ac4387d96256ca18
Signed-off-by: Alon Bar-Lev <alonbl@redhat.com>