core: sorted getJsonDiskDescription()
ImagesHandler.getJsonDiskDescription() used a HashMap to serialize a
Disk's Alias and Description. Since the order of iteration of a HashMap
cannot be trusted, it made the method's behavior unpredictable in
different environments, specifically breaking the unit tests on the
default JDK8 supplied with Fedora 21.

This patch replaces the internal implementation from using a HashMap to
use a TreeMap, ensuring the keys are iterated by order, and thus
producing a reliable output:
{"DiskAlias":"some_alias","DiskDescription":"some_description}. The unit
tests have been fixed accordingly, making them agnostic to JDK upgrades.

Change-Id: Ic8496858a2cb5715ab9f84b8ae77f62b3e6097d4
Bug-Url: https://bugzilla.redhat.com/1180522
Signed-off-by: Allon Mureinik <amureini@redhat.com>