webadmin: fix general & app list sub-tabs on import
We used to override 'setEntity' method in VmGeneralModel &
VmAppListModel classes when using them for VM which is going to be
imported, to set extract the VM from ImportVmData and use it as the
entity of the model.

Recently this part was dropped when switched to use DI in the
import-from-export-domain model, which results in class-cast exception.

This patch introduces two models, VmImportGeneralModel &
VmImportAppListModel, which inherit from VmGeneralModel &
VmAppListModel and override their 'setEntity' method as we did before,
and these models are used from the import-model instead.

Change-Id: I7235d5d6ddf953078e668b94b4d03bc136da5a77
Signed-off-by: Arik Hadas <ahadas@redhat.com>