webadmin: reverted removal of selectedItemsChanged method.
This method was mistakingly removed as a part of one big refactor.
Apparently it was removed because it seems that this method does the
same thing as overriden one, but that's not true, both calls their
private equally named method. This could be fixed by making that
method protected, but after some research this is not easily doable,
so I just returned removed method back and place warning comment to
method being overridden.

Change-Id: I473566cf0755519eb7fc922c435ae8a2ee545779
Bug-Url: https://bugzilla.redhat.com/1180897
Signed-off-by: Martin Mucha <mmucha@redhat.com>