restapi: MAC pool elements must be optional
Currently the "allow_duplicates" element of the MAC pool type of the XML
schema is mandatory. This means that links to the MAC pools fail XML
schema validation. For example, a link like the following will fail
validation because it doesn't have the mandatory "allow_duplicates"
nested element:

  <mac_pool href="..." id="..."/>

In general all elements of XML complex types that may be used as links
need to be optional.

This patch changes the XML schema so that the "allow_duplicates" element
is optional.

Change-Id: Ie35fcae56df17619bdb0464699e7a154b4765d89
Bug-Url: https://bugzilla.redhat.com/1181711
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>