webadmin: fix closing dialog and missing titles
There must have been some infra change causing all this.
Component configuring MAC pools is reused and opened in two parts of
gui. At one spot it's opened via setWindow method and in another via
setConfirmWindow, which is problematic since closing of such window is
done inside of SharedMacPoolModel, which does not know how it was
displayed. Simplest solution was to pass parameter denoting 'how to
close' when cancel is clicked.

Also dialog title wasn't set; changing model on which title is set
fixed that.

Change-Id: If38382bbe1a971f8904bbbb4fbd1db7e1855fef4
Bug-Url: https://bugzilla.redhat.com/1180898
Bug-Url: https://bugzilla.redhat.com/1180901
Signed-off-by: Martin Mucha <mmucha@redhat.com>