core: standardization of hibernate command
1. Just start to create the disks in the execute phase, no need to
update the active snapshot's memory in advance - will do it in the
end-action phase, in case we managed to create the disks.

2. Change the way we get the created disks: instead of querying the DB
in the parent command, the add-disk command returns the created disk in
its end-action phase. This is a better approach since we will need to
have the same logic in import from external provider command.

3. Cleanup the end-with-failure case: in case of a failure, just need to
make sure we clean the memory from the active snapshot and remove the
disks (revert tasks).

Change-Id: Ib7c97fb3907b155eb5d5d03b665ab80814da8229
Signed-off-by: Arik Hadas <ahadas@redhat.com>