core: lock vm for whole hibernate command life cycle
Change the VM lock which is acquired in the hibernate command such that
it will not be released at the end of the execute phase but will remain
for the whole life-cycle of the command.

Relevant message that indicates the VM is being hibernated is added to
the lock.

Currently this change does not do much as the VM is switched to
prepare-to-hibernate status anyway, but this is a step towards removing
this status.

Change-Id: I7a00eeb74338469339740e0fb20b71b8f849fc02
Signed-off-by: Arik Hadas <ahadas@redhat.com>