restapi: Add API to support warning for attached Storage Domains
Adding a REST API to support warning for attached Storage Domains on
attach or import of Storage Domain.

POST
/api/storagedomains/d97c4fc5-bf5d-41a3-94c6-8fe821f98ff0/isattached
HTTP/1.1
Accept: application/xml
Content-type: application/xml
<action>
  <host id="383a1e8d-d299-45d4-84d5-9bcc9fcf8823"></host>
</action>

The response which should returned is a Storage Domain with the flag
is_attached.
as so:
<?xml version="1.0" encoding="UTF-8" standalone="yes"?>
<storage_domain id="d97c4fc5-bf5d-41a3-94c6-8fe821f98ff0">
    <is_attached>false</is_attached>
</storage_domain>

The operation is not implemented with GET since the operation requires
the connection-host to be provided as parameter.

Change-Id: I4b12e5ea2abea66cb950e3cdd64baaaff72a6919
Bug-Url: https://bugzilla.redhat.com/1181681
Signed-off-by: Maor Lipchuk <mlipchuk@redhat.com>