core: domain might remain locked on deactivation
When passing the compensation context to the child command the
'recorded' changes aren't reverted becasue when attempting to load the
compensation data the load is by the command id which is different
between the parent command and the child command.

When no tasks are created when deactivating domain and the actual
deactivation fails, the domain remained locked because of the above
reason.

Change-Id: I4a1bea6522ceb8c83be9466ed50748559745c3c8
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1184927
Signed-off-by: Liron Aravot <laravot@redhat.com>