core: Remove redundant threshold call
While adding storage domains to a map in
ImagesHandler.fillImagesMapBasedOnTemplate(), there's a threshold check.
This check is redundant and buggy here.
Redundant because there are storage allocation checks in the system
now, which should be called in due time.

Buggy, since in case the threshold has been met, the threshold
validation fails, but with no error at the time of failure.
The map is not filled correctly, and we get another, unrelated error.

This patch removes the threshold validation and relies on it being
called when it should.

Change-Id: Ifb4357a19bc51ae5feb3c8e79ff36f3adacc327f
Bug-Url: https://bugzilla.redhat.com/1185613
Related-To: https://bugzilla.redhat.com/1179690
Signed-off-by: Vered Volansky <vvolansk@redhat.com>