webadmin: Fix running concurrent GetImagesList calls
When navigating to the Images subtab of an ISO storage domain, the
method syncSearch is called. In case one makes another similar request,
e.g., by clicking on another ISO domain, the new request gets
blocked. In this case, the ISOs belonging to the first domain are
displayed for the second domain.

This patch enables concurrent requests to be executed without being
blocked, and thus makes it possible for the new request to override the
old images list with the new one.

Change-Id: Ib6df991f597b26c13395907e99dd03934123e9b7
Bug-Url: https://bugzilla.redhat.com/1129267
Signed-off-by: Idan Shaby <ishaby@redhat.com>