webadmin: removed incorrect value change handler
The same handler has been registerd both on the ListModelTypeAheadListBox and
on BaseListModelSuggestBox (using the addValueChangeHandler).

The deleted one had no meaning - it's logic is handled by the
SuggestBoxFocusHandler but in some cases kicked in sooner than the one in
BaseListModelSuggestBox hiding the suggestions sooner than the correct handler
was able to use them.

Change-Id: I203023384e0d29671220d9ebd2ed7c4fd4b67da3
Bug-Url: https://bugzilla.redhat.com/1166022
Signed-off-by: Tomas Jelinek <tjelinek@redhat.com>