core: Leave free space value as null
In Some situations, such as a newly created disks or disks on a storage domain
that's been re-connected, the disk's free space -
StorageDomainDynamic.availableDiskSize is null. This was interpreted
into 0 in some cases, causing a strict reaction from the system, though
is doesn't necessarily means the storage has 0 space. These cases are now using the null value and
do not assume low space.
The value will be updated on the next poll from vdsm, making sure future
behaviour is as is should.

Change-Id: I3379f34c0022eb7066db400f9d9342f8d02fb85f
Bug-Url: https://bugzilla.redhat.com/1177147
Bug-Url: https://bugzilla.redhat.com/1182012
Related-To: https://bugzilla.redhat.com/1176854
Signed-off-by: Vered Volansky <vvolansk@redhat.com>