webadmin: NPE while login if there are some running tasks
The new SimpleActionTable() in the AlertsEventsFooterView's constructor
triggers the TaskModelProvider.updateDataProvider() which in case there are at
least one started job, triggers the
taskHandler.onRunningTasksCountChange(count). This in turn triggers the call to
AlertsEventsFooterView.setAlertCount() which expects the buttons to be already
inited. But this fields are inited at the end of the constructor which causes
NPE.

Fixed by calling the updateButtonResources() which inits this buttons sooner.

This bug happens in development mode only.

Change-Id: I3315802e33d0bc21ae1810802d1466464e8321b4
Signed-off-by: Tomas Jelinek <tjelinek@redhat.com>