userportal, webadmin: allow SSO robots to fill in the login form
Our UI editors/models don't get populated from forms unless a change event fires on
the form (usually on blur when a user types in a value and tabs away from that field
(ValueChangeEvent)).

For the login form, there are third-party SSO applications that "paste" credentials
into the form. LastPass, for example, works fine because it fires a DOM change event
on the fields it pastes into. But here are other solutions that don't fire that
change event after pasting into the fields.

We want to allow those to work by forcing change events on the form fields
when the form is submitted. This patch does just that. Now when the login form
is submitted, "change" is fired on the username, password, and profile fields, which
causes the editor and model to be properly populated, which allows the login to
succeed.

Change-Id: I0364ac5c4524f5260480bca38642a6998d0c9147
Bug-Url: https://bugzilla.redhat.com/1154666
Signed-off-by: Greg Sheremeta <gshereme@redhat.com>