core: Fix import from glance storage allocation check
The storage allocation check uses the
DiskImage.getActualDiskWithSnapshotsSizeInBytes() method to evaluate the
disk's size according to it's snapshots' sizes. In this case, the disk image is
not created with an active image (which is considered as a snapshot for
the aforementioned method), which resulted in a disk size of 0.
This patch adds the image to the snapshots.

Change-Id: I02d1cfb30c64a7d7b7b4c0386b134583faee190d
Bug-Url: https://bugzilla.redhat.com/1178012
Bug-Url: https://bugzilla.redhat.com/1053750
Signed-off-by: Vered Volansky <vvolansk@redhat.com>