setup: checking if command_entities table exist
oVirt 3.5 has introduced command_entities table.
In addition to adding some new functions related to this
table to the task clenaer utility, some functions got changed,
but as it is run during setup and might run on top of version which
preceeds oVirt 3.5, it should be checked in the relevant SPs that
the table exists.

Bug-Url: https://bugzilla.redhat.com/1161012
Bug-Url: https://bugzilla.redhat.com/1164771
Change-Id: I3152c4d3f6e02915053ada5019abeea7cb356751
Signed-off-by: Yair Zaslavsky <yzaslavs@redhat.com>