packaging: Fix symlink to c3p0 in Fedora 21
The c3p0 library is completely useless for us, as the engine doesn't use
it, but anyhow we have a module for it, and the symlink to the .jar file
is broken in Fedora 21, as it has been moved from "/usr/share/java" to
"/usr/share/java/c3p0". This causes a failure during the startup of the
application server. This patch fixes the RPM spec so that it creates the
correct symlink in Fedora 21.

Change-Id: Iebc360999c5b6beaf1c60722f1c3003b95f27ec4
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>