restapi: Calculate singulars correctly
The RSDL generator doesn't calculate singulars correctly. For example,
for the plural form "schedulingpolicies" it calculats the singular as
"schedulingpolicie". This is incorrect and doesn't match with the
singular form specified in the RSDL metadata. As a result the parameter
metadata for some entities and collections, in particular for scheduling
policy related entities and collections, isn't generated. As a result of
that the Python SDK doesn't contain the metadata, and thus the CLI
auto-completion doesn't work. This patch fixes the RSDL generator so
that singulars will be calculated correctly.

Change-Id: Ifbd00642de18637e2620433395b3065a062a4ee9
Bug-Url: https://bugzilla.redhat.com/1189095
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>