restapi: Extending disk editing when adding a new template.
Currently when adding a new template, its disks edit is enabled only
when the disk's storage domain is specified- In this case, editing
any other field is ignored by the REST, and the request which is sent
to the backend does not contain all the values the user posted.

This patch makes the REST request for a new template to consider every
given value related to the new template's disks whether the storage
domain is specified and whether it is not, so the disk values that
are provided will always be sent to backend.

The specified attributes given in the REST request will be patched on
the original disk using a mapper so that a sent disk will contain a
default value in case some of its fields will not be provided by the
user.

Change-Id: Ib0baf4e5bd7233fbc4eab2f1f671b8b15e08ba03
Bug-Url: https://bugzilla.redhat.com/1110798
Signed-off-by: Amit Aviram <aaviram@redhat.com>