engine : Refactoring in accordance with vdsm patches
This patch does the following :

1. The geo-rep engine patches were merged before the corresponding
patches and hence at merge time, contained the exception codes and
api names as the vdsm exposed at that time.
       Now, that the vdsm patches have been merged, this patch
refactors the engine side accordingly.
2. StartGlusterVolumeGeoRepCommand was modifying the status of geo-rep
session in db on success. But, it is observed that the session on gluster
side takes a while to change state to ACTIVE untill then it stays as
INITIALISING and hence this patch removes the state change code and allows
the geo-rep sync job to detect state change.

Change-Id: I702d6c62f576cf2a49437fcf37155290d8580588
Signed-off-by: Anmol Babu <anbabu@redhat.com>