webadmin: fix cpu-profile setting on import dialog
In the new two-phases import dialog, the cpu-profile setting was cleared
when switching to the second screen and then back to the first screen.

This patch fix it by not initializing a new instance of
ImportVmFromExportDomainModel on every switch to the second screen.
Instead, we initialize it in advance and use the same instance.

Change-Id: I8bb3e500952949acf3795a5f92dd6160a4df9171
Signed-off-by: Arik Hadas <ahadas@redhat.com>