restapi: Remove ReadOnlyResources interface
The presence of this interface induces errors in the RSDL generator. In
particular the type of the disk snapshots collection is reported as
"BaseResource" instead of "DiskSnapshot", which causes additional issues
in the consumers of the RSDL, including the generators of the Python and
Java SDKs. To avoid this issue this patch removes completely that
interface.

Change-Id: Id449ea7bb88f7acae0a3b536aa2d45e1ddbeb48e
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>