restapi: Document All-Content header for hosts
The "All-Content" header is used by the hosts collection and by the host
resource, but it isn't documented in the RSDL metadata. As a result it
isn't described in the RSDL and it isn't available in the SDKs or the
CLI. This patch fixes the RSDL metadata.

Change-Id: I880cd2f00a74813defd8366285fefc9c6c1bacf6
Bug-Url: https://bugzilla.redhat.com/1194185
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>
(cherry picked from commit afae793f20adb5cbaf9da8df302251723091ea49)