webadmin: StorageListModel: refactoring storage prepare for edit.
As preparing storage for editing in StorageListModel has similiar logic
through File and Block storage and all of their types, and currently
there is a "prepare" method for each of those types- it is now
consolidated into each model, with the proper differences divided
between the models.

prepareStorageForEdit will replace all of these "prepareXstorageForEdit"
and let the proper model to do the rest of the work, divided through SAN and
FILE storage types.

Change-Id: I2bb2a4c142eb1fe2870dc00d0d8764a6c866f8a2
Signed-off-by: Amit Aviram <aaviram@redhat.com>