restapi: made sure only the supported template entities are copied
The blank template is now editable. It is not connected to any cluster (similar
to instance types) and it supportrs all options from the highest cluster
version.

When a VM is created from this template in the older cluster, it is
important not to copy fields which are not supported on the target cluster.

This is not relevant for the VM pools since the pools can be created only from
non-blan templates.

Feature Page: www.ovirt.org/Features/Blank_to_Defaults
Change-Id: I7430bca85b95ddf97eaa93abbcdf13885e017f9e
Bug-Url: https://bugzilla.redhat.com/1130174
Signed-off-by: Tomas Jelinek <tjelinek@redhat.com>