core, engine: servlet to support the console proxy
This patch creates two services that provide information to the SSH
server that acts as the VirtIO console proxy in the
/services/consoleproxy endpoint in the engine, this endpoint
accepts two commands:

- "public_keys": Provides to the authkeys script the list of
  allowed public keys into the SSH server and the respective ID of
  each allowed user.

- "available_consoles": Lists the available VMs to be accessed
by the user given his ID and in which host these VMs are located.

In order to protect these services, a ticket must emitted with the
desired command and arguments by the client to the service.

Feature wiki page: http://www.ovirt.org/Features/Serial_Console

Bug-Url: https://bugzilla.redhat.com/1223671
Change-Id: I53c721da21cefcf4069d14c7016b6f7d97f9eac9
Signed-off-by: Vitor de Lima <vdelima@redhat.com>
Signed-off-by: Francesco Romani <fromani@redhat.com>