db: reformat views code using pgformatter
Currently it is hard to review changes made in DB scripts
The reason is lack of general formatting.
For example, in views, there are many columns in the same row and when a
column is added it is hard to track the change. That may lead to some
problems and regressions after the changed code is merged. The following
uses a opensource tool named pgFormatter
http://sourceforge.net/projects/pgformatter/ to make our code more
readable and friendly for reviews.

Currently only engine and dwh views were changed

Change-Id: I8c127f152c73f08ac8d7d6f9448d010235e9d3d2
Signed-off-by: emesika <emesika@redhat.com>