core: determine whether domain monitoring result is actual
When the domain monitoring results are reported back from vdsm, the
first monitor run may haven't finished yet. In that case the returned
status from vdsm is a dummy status and not an actual result.
The non actual status could lead to a wrong decision making on the engine
status (for example - decide that an inactive domain is now active).

This patch uses the logic added on vdsm side (change I1fea5) to inspect
whether the reported status is actual or not.
For earlier version that doesn't return that info, each returned report
will be considered as actual to leave the behavior as it was.

Change-Id: Iddbba94d81694bd6cb1ce3648872eb81182a50ab
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1183977
Signed-off-by: laravot@redhat.com <laravot@redhat.com>