engine: Correct user connected/disconnected messages
Message 'User disconnected' was added to Action Log when user had
connected to a VM and vice versa.

The bug was in VdsEventListener.processOnClientIpChange() method called
from VmsMonitoring.afterVMsRefreshTreatment() method.
processOnClientIpChange() was reading the client IP from DB, but IP
address of the currently connecting client is written to DB after
calling this method (saveVmsToDb() call at the end of
afterVMsRefreshTreatment() method).

To correct this additional parameter was added to
processOnClientIpChange() method to pass IP address of the currently
connecting client.

Change-Id: Icce32619e3cc88f12e339b4eef5717f3d279e232
Signed-off-by: Shmuel Melamud <smelamud@redhat.com>