restapi,core: Enabling disk's description edit when adding a new template.
Currently when adding a new template the user is able to change the
template's disk's alias.
Editing the disk's description looks natural to edit to the user
along editing the alias, and also seems to be a relevant use case
when adding a new template.
This patch adds the possibility to also edit the disk's description.

Change-Id: Ic68114c072565ad2d0938fc0256448da9bce0910
Signed-off-by: Amit Aviram <aaviram@redhat.com>