core: query long-time locks
Add a method in LockManager that returns the long-time locks for a given
key. By "long-time lock" we refer to locks that are used in
async-commands for their whole execution.

This patch contains only the infrastructure. This change will be used by
queries (searches) from the UI, so entities will be returned with their
locks state, so it could be presented in the UI.

Change-Id: Iafcab44ef52d21405eeaaeb2cda4c03864a0e8e8
Signed-off-by: Arik Hadas <ahadas@redhat.com>