webadmin: Merged the status column with migration progress
Main tab virtual machine the migration progress has been removed and merged
with the status. The "status" column now contains the status name as before if
the VM is not migrating. But if the VM is migrating, the status column contains
the migration progress bar and inside it the status name and percentage (e.g.
"Migraing from: 30%" inside a progress bar).

The MigratingTo status is not taken into account because the VM contains the
migration progress only when the status is MigratingFrom.

Change-Id: I8794891e5b2c6e8eb1339dd24a68fee0ba1695f1
Signed-off-by: Tomas Jelinek <tjelinek@redhat.com>